backend/requirements.txt
fastapi
uvicorn
sqlalchemy
sqlmodel
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings
requests
httpx
beautifulsoup4
pypdf2
python-docx
sentence-transformers
faiss-cpu
langchain
langchain-community
jinja2
python-dotenv
websockets
reportlab
google-generativeai
gtts
